+++ big.js
@@ -1,8 +1,8 @@
 /*
- *  big.js v5.2.2
+ *  big.js v6.0.0
  *  A small, fast, easy-to-use library for arbitrary-precision decimal arithmetic.
- *  Copyright (c) 2018 Michael Mclaughlin <M8ch88l@gmail.com>
- *  https://github.com/MikeMcl/big.js/LICENCE
+ *  Copyright (c) 2020 Michael Mclaughlin
+ *  https://github.com/MikeMcl/big.js/LICENCE.md
  */
 ;(function (GLOBAL) {
   'use strict';
@@ -18,7 +18,7 @@
      * The maximum number of decimal places (DP) of the results of operations involving division:
      * div and sqrt, and pow with negative exponents.
      */
-    DP = 20,          // 0 to MAX_DP
+    DP = 20,            // 0 to MAX_DP
 
     /*
      * The rounding mode (RM) used when rounding to the above decimal places.
@@ -46,10 +46,16 @@
     /*
      * The positive exponent (PE) at and above which toString returns exponential notation.
      * (JavaScript numbers: 21)
-     * 1000000 is the maximum recommended exponent value of a Big.
-     * (This limit is not enforced or checked.)
+     * 1000000 is the maximum recommended exponent value of a Big, but this limit is not enforced.
      */
     PE = 21,            // 0 to 1000000
+
+    /*
+     * When true, an error will be thrown if a primitive number is passed to the Big constructor,
+     * or if valueOf is called, or if toNumber is called on a Big which cannot be converted to a
+     * primitive number without a loss of precision.
+     */
+    STRICT = false,     // true or false
 
 
 /**************************************************************************************************/
@@ -70,7 +76,6 @@
 
   /*
    * Create and return a Big constructor.
-   *
    */
   function _Big_() {
 
@@ -92,13 +97,20 @@
         x.e = n.e;
         x.c = n.c.slice();
       } else {
+        if (typeof n !== 'string') {
+          if (Big.strict === true) {
+            throw TypeError(INVALID + 'number');
+          }
+
+          // Minus zero?
+          n = n === 0 && 1 / n < 0 ? '-0' : String(n);
+        }
+
         parse(x, n);
       }
 
-      /*
-       * Retain a reference to this Big constructor, and shadow Big.prototype.constructor which
-       * points to Object.
-       */
+      // Retain a reference to this Big constructor.
+      // Shadow Big.prototype.constructor which points to Object.
       x.constructor = Big;
     }
 
@@ -107,7 +119,7 @@
     Big.RM = RM;
     Big.NE = NE;
     Big.PE = PE;
-    Big.version = '5.2.2';
+    Big.strict = STRICT;
 
     return Big;
   }
@@ -122,9 +134,9 @@
   function parse(x, n) {
     var e, i, nl;
 
-    // Minus zero?
-    if (n === 0 && 1 / n < 0) n = '-0';
-    else if (!NUMERIC.test(n += '')) throw Error(INVALID + 'number');
+    if (!NUMERIC.test(n)) {
+      throw Error(INVALID + 'number');
+    }
 
     // Determine sign.
     x.s = n.charAt(0) == '-' ? (n = n.slice(1), -1) : 1;
@@ -170,69 +182,67 @@
 
 
   /*
-   * Round Big x to a maximum of dp decimal places using rounding mode rm.
-   * Called by stringify, P.div, P.round and P.sqrt.
+   * Round Big x to a maximum of sd significant digits using rounding mode rm.
    *
    * x {Big} The Big to round.
-   * dp {number} Integer, 0 to MAX_DP inclusive.
-   * rm {number} 0, 1, 2 or 3 (DOWN, HALF_UP, HALF_EVEN, UP)
+   * sd {number} Significant digits: integer, 0 to MAX_DP inclusive.
+   * rm {number} Rounding mode: 0 (down), 1 (half-up), 2 (half-even) or 3 (up).
    * [more] {boolean} Whether the result of division was truncated.
    */
-  function round(x, dp, rm, more) {
-    var xc = x.c,
-      i = x.e + dp + 1;
+  function round(x, sd, rm, more) {
+    var xc = x.c;
 
-    if (i < xc.length) {
-      if (rm === 1) {
+    if (rm === UNDEFINED) rm = Big.RM;
+    if (rm !== 0 && rm !== 1 && rm !== 2 && rm !== 3) {
+      throw Error(INVALID_RM);
+    }
 
-        // xc[i] is the digit after the digit that may be rounded up.
-        more = xc[i] >= 5;
-      } else if (rm === 2) {
-        more = xc[i] > 5 || xc[i] == 5 &&
-          (more || i < 0 || xc[i + 1] !== UNDEFINED || xc[i - 1] & 1);
-      } else if (rm === 3) {
-        more = more || !!xc[0];
+    if (sd < 1) {
+      more =
+        rm === 3 && (more || !!xc[0]) || sd === 0 && (
+        rm === 1 && xc[0] >= 5 ||
+        rm === 2 && (xc[0] > 5 || xc[0] === 5 && (more || xc[1] !== UNDEFINED))
+      );
+
+      xc.length = 1;
+
+      if (more) {
+
+        // 1, 0.1, 0.01, 0.001, 0.0001 etc.
+        x.e = x.e - sd + 1;
+        xc[0] = 1;
       } else {
-        more = false;
-        if (rm !== 0) throw Error(INVALID_RM);
+
+        // Zero.
+        xc[0] = x.e = 0;
       }
+    } else if (sd < xc.length) {
 
-      if (i < 1) {
-        xc.length = 1;
+      // xc[sd] is the digit after the digit that may be rounded up.
+      more =
+        rm === 1 && xc[sd] >= 5 ||
+        rm === 2 && (xc[sd] > 5 || xc[sd] === 5 &&
+          (more || xc[sd + 1] !== UNDEFINED || xc[sd - 1] & 1)) ||
+        rm === 3 && (more || !!xc[0]);
 
-        if (more) {
+      // Remove any digits after the required precision.
+      xc.length = sd--;
 
-          // 1, 0.1, 0.01, 0.001, 0.0001 etc.
-          x.e = -dp;
-          xc[0] = 1;
-        } else {
+      // Round up?
+      if (more) {
 
-          // Zero.
-          xc[0] = x.e = 0;
-        }
-      } else {
-
-        // Remove any digits after the required decimal places.
-        xc.length = i--;
-
-        // Round up?
-        if (more) {
-
-          // Rounding up may mean the previous digit has to be rounded up.
-          for (; ++xc[i] > 9;) {
-            xc[i] = 0;
-            if (!i--) {
-              ++x.e;
-              xc.unshift(1);
-            }
+        // Rounding up may mean the previous digit has to be rounded up.
+        for (; ++xc[sd] > 9;) {
+          xc[sd] = 0;
+          if (!sd--) {
+            ++x.e;
+            xc.unshift(1);
           }
         }
-
-        // Remove trailing zeros.
-        for (i = xc.length; !xc[--i];) xc.pop();
       }
-    } else if (rm < 0 || rm > 3 || rm !== ~~rm) {
-      throw Error(INVALID_RM);
+
+      // Remove trailing zeros.
+      for (sd = xc.length; !xc[--sd];) xc.pop();
     }
 
     return x;
@@ -242,48 +252,14 @@
   /*
    * Return a string representing the value of Big x in normal or exponential notation.
    * Handles P.toExponential, P.toFixed, P.toJSON, P.toPrecision, P.toString and P.valueOf.
-   *
-   * x {Big}
-   * id? {number} Caller id.
-   *         1 toExponential
-   *         2 toFixed
-   *         3 toPrecision
-   *         4 valueOf
-   * n? {number|undefined} Caller's argument.
-   * k? {number|undefined}
-   * rm? 0, 1, 2 or 3 (ROUND_DOWN, ROUND_HALF_UP, ROUND_HALF_EVEN, ROUND_UP)
    */
-  function stringify(x, id, n, k, rm) {
-    var e, s,
-      Big = x.constructor,
-      z = !x.c[0];
-
-    if (n !== UNDEFINED) {
-      if (n !== ~~n || n < (id == 3) || n > MAX_DP) {
-        throw Error(id == 3 ? INVALID + 'precision' : INVALID_DP);
-      }
-
-      x = new Big(x);
-
-      // The index of the digit that may be rounded up.
-      n = k - x.e;
-
-      // Round?
-      if (x.c.length > ++k) round(x, n, rm === undefined ? Big.RM : rm);
-
-      // toFixed: recalculate k as x.e may have changed if value rounded up.
-      if (id == 2) k = x.e + n + 1;
-
-      // Append zeros?
-      for (; x.c.length < k;) x.c.push(0);
-    }
-
-    e = x.e;
-    s = x.c.join('');
-    n = s.length;
+  function stringify(x, doExponential, isNonzero) {
+    var e = x.e,
+      s = x.c.join(''),
+      n = s.length;
 
     // Exponential notation?
-    if (id != 2 && (id == 1 || id == 3 && k <= e || e <= Big.NE || e >= Big.PE)) {
+    if (doExponential) {
       s = s.charAt(0) + (n > 1 ? '.' + s.slice(1) : '') + (e < 0 ? 'e' : 'e+') + e;
 
     // Normal notation.
@@ -291,13 +267,16 @@
       for (; ++e;) s = '0' + s;
       s = '0.' + s;
     } else if (e > 0) {
-      if (++e > n) for (e -= n; e--;) s += '0';
-      else if (e < n) s = s.slice(0, e) + '.' + s.slice(e);
+      if (++e > n) {
+        for (e -= n; e--;) s += '0';
+      } else if (e < n) {
+        s = s.slice(0, e) + '.' + s.slice(e);
+      }
     } else if (n > 1) {
       s = s.charAt(0) + '.' + s.slice(1);
     }
 
-    return x.s < 0 && (!z || id == 4) ? '-' + s : s;
+    return x.s < 0 && isNonzero ? '-' + s : s;
   }
 
 
@@ -318,7 +297,7 @@
    * Return 1 if the value of this Big is greater than the value of Big y,
    *       -1 if the value of this Big is less than the value of Big y, or
    *        0 if they have the same value.
-  */
+   */
   P.cmp = function (y) {
     var isneg,
       x = this,
@@ -364,10 +343,14 @@
       k = x.s == y.s ? 1 : -1,
       dp = Big.DP;
 
-    if (dp !== ~~dp || dp < 0 || dp > MAX_DP) throw Error(INVALID_DP);
+    if (dp !== ~~dp || dp < 0 || dp > MAX_DP) {
+      throw Error(INVALID_DP);
+    }
 
     // Divisor is zero?
-    if (!b[0]) throw Error(DIV_BY_ZERO);
+    if (!b[0]) {
+      throw Error(DIV_BY_ZERO);
+    }
 
     // Dividend is 0? Return +-0.
     if (!a[0]) return new Big(k * 0);
@@ -381,10 +364,10 @@
       q = y,                // quotient
       qc = q.c = [],
       qi = 0,
-      d = dp + (q.e = x.e - y.e) + 1;    // number of digits of the result
+      p = dp + (q.e = x.e - y.e) + 1;    // precision of the result
 
     q.s = k;
-    k = d < 0 ? 0 : d;
+    k = p < 0 ? 0 : p;
 
     // Create version of divisor with leading zero.
     bz.unshift(0);
@@ -445,10 +428,11 @@
       // There can't be more than one zero.
       qc.shift();
       q.e--;
+      p--;
     }
 
     // Round?
-    if (qi > d) round(q, dp, Big.RM, r[0] !== UNDEFINED);
+    if (qi > p) round(q, p, Big.RM, r[0] !== UNDEFINED);
 
     return q;
   };
@@ -458,7 +442,7 @@
    * Return true if the value of this Big is equal to the value of Big y, otherwise return false.
    */
   P.eq = function (y) {
-    return !this.cmp(y);
+    return this.cmp(y) === 0;
   };
 
 
@@ -612,7 +596,9 @@
       a = x.s,
       b = (y = new Big(y)).s;
 
-    if (!y.c[0]) throw Error(DIV_BY_ZERO);
+    if (!y.c[0]) {
+      throw Error(DIV_BY_ZERO);
+    }
 
     x.s = y.s = 1;
     ygtx = y.cmp(x) == 1;
@@ -716,7 +702,10 @@
       y = one,
       isneg = n < 0;
 
-    if (n !== ~~n || n < -MAX_POWER || n > MAX_POWER) throw Error(INVALID + 'exponent');
+    if (n !== ~~n || n < -MAX_POWER || n > MAX_POWER) {
+      throw Error(INVALID + 'exponent');
+    }
+
     if (isneg) n = -n;
 
     for (;;) {
@@ -731,20 +720,35 @@
 
 
   /*
-   * Return a new Big whose value is the value of this Big rounded using rounding mode rm
-   * to a maximum of dp decimal places, or, if dp is negative, to an integer which is a
-   * multiple of 10**-dp.
+   * Return a new Big whose value is the value of this Big rounded to a maximum precision of sd
+   * significant digits using rounding mode rm, or Big.RM if rm is not specified.
+   *
+   * sd {number} Significant digits: integer, 1 to MAX_DP inclusive.
+   * rm? {number} Rounding mode: 0 (down), 1 (half-up), 2 (half-even) or 3 (up).
+   */
+  P.prec = function (sd, rm) {
+    if (sd !== ~~sd || sd < 1 || sd > MAX_DP) {
+      throw Error(INVALID + 'precision');
+    }
+    return round(new this.constructor(this), sd, rm);
+  };
+
+
+  /*
+   * Return a new Big whose value is the value of this Big rounded to a maximum of dp decimal places
+   * using rounding mode rm, or Big.RM if rm is not specified.
+   * If dp is negative, round to an integer which is a multiple of 10**-dp.
    * If dp is not specified, round to 0 decimal places.
-   * If rm is not specified, use Big.RM.
    *
    * dp? {number} Integer, -MAX_DP to MAX_DP inclusive.
-   * rm? 0, 1, 2 or 3 (ROUND_DOWN, ROUND_HALF_UP, ROUND_HALF_EVEN, ROUND_UP)
+   * rm? {number} Rounding mode: 0 (down), 1 (half-up), 2 (half-even) or 3 (up).
    */
   P.round = function (dp, rm) {
-    var Big = this.constructor;
     if (dp === UNDEFINED) dp = 0;
-    else if (dp !== ~~dp || dp < -MAX_DP || dp > MAX_DP) throw Error(INVALID_DP);
-    return round(new Big(this), dp, rm === UNDEFINED ? Big.RM : rm);
+    else if (dp !== ~~dp || dp < -MAX_DP || dp > MAX_DP) {
+      throw Error(INVALID_DP);
+    }
+    return round(new this.constructor(this), dp + this.e + 1, rm);
   };
 
 
@@ -764,7 +768,9 @@
     if (!x.c[0]) return new Big(x);
 
     // Negative?
-    if (s < 0) throw Error(NAME + 'No square root');
+    if (s < 0) {
+      throw Error(NAME + 'No square root');
+    }
 
     // Estimate.
     s = Math.sqrt(x + '');
@@ -789,7 +795,7 @@
       r = half.times(t.plus(x.div(t)));
     } while (t.c.slice(0, e).join('') !== r.c.slice(0, e).join(''));
 
-    return round(r, Big.DP -= 4, Big.RM);
+    return round(r, (Big.DP -= 4) + r.e + 1, Big.RM);
   };
 
 
@@ -862,42 +868,100 @@
 
 
   /*
-   * Return a string representing the value of this Big in exponential notation to dp fixed decimal
-   * places and rounded using Big.RM.
+   * Return a string representing the value of this Big in exponential notation rounded to dp fixed
+   * decimal places using rounding mode rm, or Big.RM if rm is not specified.
    *
-   * dp? {number} Integer, 0 to MAX_DP inclusive.
-   * rm? 0, 1, 2 or 3 (ROUND_DOWN, ROUND_HALF_UP, ROUND_HALF_EVEN, ROUND_UP)
+   * dp? {number} Decimal places: integer, 0 to MAX_DP inclusive.
+   * rm? {number} Rounding mode: 0 (down), 1 (half-up), 2 (half-even) or 3 (up).
    */
   P.toExponential = function (dp, rm) {
-    return stringify(this, 1, dp, dp, rm);
+    var x = this,
+      n = x.c[0];
+
+    if (dp !== UNDEFINED) {
+      if (dp !== ~~dp || dp < 0 || dp > MAX_DP) {
+        throw Error(INVALID_DP);
+      }
+      x = round(new x.constructor(x), ++dp, rm);
+      for (; x.c.length < dp;) x.c.push(0);
+    }
+
+    return stringify(x, true, !!n);
   };
 
 
   /*
-   * Return a string representing the value of this Big in normal notation to dp fixed decimal
-   * places and rounded using Big.RM.
+   * Return a string representing the value of this Big in normal notation rounded to dp fixed
+   * decimal places using rounding mode rm, or Big.RM if rm is not specified.
    *
-   * dp? {number} Integer, 0 to MAX_DP inclusive.
-   * rm? 0, 1, 2 or 3 (ROUND_DOWN, ROUND_HALF_UP, ROUND_HALF_EVEN, ROUND_UP)
-   * 
+   * dp? {number} Decimal places: integer, 0 to MAX_DP inclusive.
+   * rm? {number} Rounding mode: 0 (down), 1 (half-up), 2 (half-even) or 3 (up).
+   *
    * (-0).toFixed(0) is '0', but (-0.1).toFixed(0) is '-0'.
    * (-0).toFixed(1) is '0.0', but (-0.01).toFixed(1) is '-0.0'.
    */
   P.toFixed = function (dp, rm) {
-    return stringify(this, 2, dp, this.e + dp, rm);
+    var x = this,
+      n = x.c[0];
+
+    if (dp !== UNDEFINED) {
+      if (dp !== ~~dp || dp < 0 || dp > MAX_DP) {
+        throw Error(INVALID_DP);
+      }
+      x = round(new x.constructor(x), dp + x.e + 1, rm);
+
+      // x.e may have changed if the value is rounded up.
+      for (dp = dp + x.e + 1; x.c.length < dp;) x.c.push(0);
+    }
+
+    return stringify(x, false, !!n);
+  };
+
+
+  /*
+   * Return a string representing the value of this Big in exponential notation.
+   * Include the sign for negative zero.
+   */
+  P.toJSON = function () {
+    return stringify(this, true, true);
+  };
+
+
+  /*
+   * Return the value of this Big as a primitve number.
+   */
+  P.toNumber = function () {
+    var n = Number(stringify(this, true, true));
+    if (this.constructor.strict === true && !this.eq(n.toString())) {
+      throw Error(NAME + 'Imprecise conversion');
+    }
+    return n;
   };
 
 
   /*
    * Return a string representing the value of this Big rounded to sd significant digits using
-   * Big.RM. Use exponential notation if sd is less than the number of digits necessary to represent
+   * rounding mode rm, or Big.RM if rm is not specified.
+   * Use exponential notation if sd is less than the number of digits necessary to represent
    * the integer part of the value in normal notation.
    *
-   * sd {number} Integer, 1 to MAX_DP inclusive.
-   * rm? 0, 1, 2 or 3 (ROUND_DOWN, ROUND_HALF_UP, ROUND_HALF_EVEN, ROUND_UP)
+   * sd {number} Significant digits: integer, 1 to MAX_DP inclusive.
+   * rm? {number} Rounding mode: 0 (down), 1 (half-up), 2 (half-even) or 3 (up).
    */
   P.toPrecision = function (sd, rm) {
-    return stringify(this, 3, sd, sd - 1, rm);
+    var x = this,
+      Big = x.constructor,
+      n = x.c[0];
+
+    if (sd !== UNDEFINED) {
+      if (sd !== ~~sd || sd < 1 || sd > MAX_DP) {
+        throw Error(INVALID + 'precision');
+      }
+      x = round(new Big(x), sd, rm);
+      for (; x.c.length < sd;) x.c.push(0);
+    }
+
+    return stringify(x, sd <= x.e || x.e <= Big.NE || x.e >= Big.PE, !!n);
   };
 
 
@@ -908,7 +972,9 @@
    * Omit the sign for negative zero.
    */
   P.toString = function () {
-    return stringify(this);
+    var x = this,
+      Big = x.constructor;
+    return stringify(x, x.e <= Big.NE || x.e >= Big.PE, !!x.c[0]);
   };
 
 
@@ -918,8 +984,13 @@
    * Big.PE, or a negative exponent equal to or less than Big.NE.
    * Include the sign for negative zero.
    */
-  P.valueOf = P.toJSON = function () {
-    return stringify(this, 4);
+  P.valueOf = function () {
+    var x = this,
+      Big = x.constructor;
+    if (Big.strict === true) {
+      throw Error(NAME + 'valueOf disallowed');
+    }
+    return stringify(x, x.e <= Big.NE || x.e >= Big.PE, true);
   };